backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
confluent-kafka==2.3.0
redis==5.0.1
celery==5.3.4
scikit-learn==1.4.0
google-generativeai==0.3.2
requests==2.31.0
jinja2==3.1.3
aiofiles==23.2.1
websockets==12.0
